Add a -stdin option to vgen and remove the need for vgen_stdin
Currently, the script can build either with stdin support for the UART
(`vgen_stdin`) or with tracing `vgen -trace` but it can't do both in the
same build.  The difference between the two builds is only the main .c
file, so move that into vgen.  `vgen_stdin` is now a tiny wrapper that
accepts all of the same arguments as vgen by simply forwarding to vgen.